requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
asgiref==3.5.2
async-timeout==4.0.3
asynctest==0.13.0
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
Django==3.2.15
django-cors-headers==4.1.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
flit_core==3.6.0
frozenlist==1.3.3
idna==3.4
importlib-metadata==6.7.0
multidict==6.0.4
openai==0.28.1
Pillow==9.5.0
pip==22.3.1
psycopg2==2.9.3
PyJWT==1.7.1
python-decouple==3.8
pytz==2022.7
requests==2.31.0
setuptools==65.6.3
sqlparse==0.4.3
tqdm==4.66.1
typing_extensions==4.4.0
urllib3==2.0.5
wheel==0.38.4
yarl==1.9.2
zipp==3.15.0
